avoid most of warnings

+++ user/testptelibrary.c
@@ -12,7 +12,7 @@ void umain(int argc, char **argv) {
         childofspawn();
     }
 
-    if ((r = syscall_mem_alloc(0, TMPVA, PTE_R | PTE_V | PTE_LIBRARY)) < 0)
+    if ((r = syscall_mem_alloc(0, (u_int)TMPVA, (u_int)(PTE_R | PTE_V | PTE_LIBRARY))) < 0)
         user_panic("syscall error!");
     if ((r = fork()) < 0)
         user_panic("fork: %d", r);
